backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
supabase==1.0.0
python-dotenv==1.0.0
pydantic>=1.10.7,<2.0.0
httpx>=0.23.0,<0.24.0
langgraph>=0.1.19,<1.1.0
openai>=1.0.0